requirements.txt
langchain==0.3
fastapi
uvicorn[standard]
openai
chromadb
aiohttp
pydantic
python-dotenv
pydantic-settings